fix(proxy): add condition to decide whether to perform circumvention
Previously, circumvention was skipped only when patternMatched was
false. This logic has been expanded to also skip circumvention when
the destination IP address is private, as this is unnecessary for
destinations under our control.

Consequently, an isPrivate function has been added to the proxy module
to check if any of the IP addresses from a DNS lookup are private.