fix: condition for shouldExploit (#306)
* version: prepare for 1.0.3-dev

* ci: add checks for new PR

* fix(proxy): add condition to decide whether to perform circumvention

Previously, circumvention was skipped only when patternMatched was
false. This logic has been expanded to also skip circumvention when
the destination IP address is private, as this is unnecessary for
destinations under our control.

Consequently, an isPrivate function has been added to the proxy module
to check if any of the IP addresses from a DNS lookup are private.

* ci: use golangci-lint-action@v8

* fix: rules for golangci

* ci: add libpcap dependency for lint action

* ci: cache libpcap for golangci

* ci: use golangci/golangci-lint-action@v8

* ci: add awalsh128/cache-apt-pkgs-action for golangci

* ci: rename 'Go Lint' to 'Lint'

* ci: add ignore labels for change logs in release drafter

* ci: use setup-go@v6